requirements.txt
django==3.1.5
django-phonenumber-field
django-phonenumber-field[phonenumbers]